+++ web/src/interfaces/IHeater.ts
@@ -1,0 +1,9 @@
+export interface IHeater {
+  id: number;
+  name: string;
+  preference:number;
+  pinNr:number;
+  watt:number;
+  useForMash:boolean;
+  useForBoil:boolean;
+}