+++ web/src/interfaces/ISystemSettings.ts
@@ -2,8 +2,6 @@ import TemperatureScale from '@/enums/TemperatureScale';
 
 export interface ISystemSettings {
   onewirePin: number;
-  heat1Pin: number;
-  heat2Pin: number;
   stirPin: number;
   invertOutputs: boolean;
   mqttUri: string;