+++ web/src/views/Control.vue
@@ -481,7 +481,7 @@ onBeforeUnmount(() => {
       </v-row>
       <v-row>
         <v-col cols="12" md="3">
-          <v-select label="Mash Schedule" v-model="selectedMashSchedule" :items="mashSchedules" item-title="name" :filled="mashSchedules" clearable return-object />
+          <v-select label="Mash/Boil Schedule" v-model="selectedMashSchedule" :items="mashSchedules" item-title="name" :filled="mashSchedules" clearable return-object />
         </v-col>
         <v-col cols="12" md="3">
           <v-btn color="success" class="mt-4" block @click="start"> Start </v-btn>