+++ Changelog.md
@@ -1,4 +1,10 @@
 
+# Version 1.3.0
+- Heaters are now fully dynamic, up to 10!.
+- Configurable wattage per heater.
+- Heaters can now be enabled/disabled for boil/mash.
+- Schedules have a new flag to set then as boil/mash.
+
 # Version 1.2.0
 - Support for Celsius and Fahrenheit.
 - PID loopt time is now configurable.
@@ -28,7 +34,7 @@
 - Temperature sensor settings have been significantly changed, so it is recommended to do a flash_erase and reflash
 
 ```bash
-idf.py flash_erase
+idf.py flash-erase
 ``` 
 
 - Flash Loader and firmware a normal.